# alerta5DBIO/public/json/apidocs.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "API JSON de consulta y gestión de la base de datos hidrometeorológica alerta5",
        "description": "sim: funciones CRUD sobre los registros de simulación. accessors: funciones de acceso a recursos externos y actualización de la db"
    },
    "servers": [
        {
            "url": "",
            "description": "local deployment for testing"
        }
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
   "paths": {
        "/login": {
            "post": {
                "summary": "login endpoint",
                "nickname": "login",
                "parameters": [
                    {
                        "name": "username",
                        "required": true,
                        "type": "string",
                        "in": "query"
                    },
                    {
                        "name": "password",
                        "required": true,
                        "type": "string",
                        "format": "password",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "login correct"
                    },
                    "400": {
                        "description": "login incorrect"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "summary": "logout endpoint",
                "nickname": "logout",
                "responses": {
                    "200": {
                        "description": "logout correct"
                    },
                    "400": {
                        "description": "logout incorrect"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "description": "cookie provista por el servidor en /login",
                "name": "id",
                "in": "cookie"
            }
        }
    }
}